Added "Mount and Open Finder.rb" appscript
`Mount and Open Finder.rb` is a modification of the `Open Transmit
Here.rb` script.

Instead of opening the current Finder location in a remote browser
window in Transmit, the remote disk is mounted, and a Finder window is
targeted at the chosen location.

Current hard-coded Finder window options are:

* Toolbar is hidden

* List view is selected

* Folder items are sorted by Date Modified